refactor: streamline running on localnet
The main goal: if you're already running a local NEAR network, make it
as easy as possible to run a local contract-helper & Wallet on that
network.

Someone in such a situation will very likely be running a different app
using Parcel on port 1234, and will need to sign into Wallet using a
different port. This switches the local Wallet port to 4000.

There is no need to gitignore `.env.local` since it stores no sensitive
info. Including it makes it much easier to run on a local network with
one command:

    NODE_ENV=local yarn start

Finally, note that when running a local network with nearup or
rainbow-bridge-cli, the only validator account is named `node0`. Using
this as the account creator in contract-helper is the simplest way to
get started (see https://github.com/near/near-contract-helper/pull/221),
which means the account suffix here needs to match.

I think this is a confusing suffix and would much prefer `localnet`, but
I don't know if that's worth doing now. Ideally such a change would flow
out from `nearup`, rather than being part of the setup in
`near-contract-helper` or here.